--- RecoBTag/SoftLepton/interface/MuonTaggerNoIP.h
@@ -1,39 +1,0 @@
-#ifndef RecoBTag_SoftLepton_MuonTaggerNoIP_h
-#define RecoBTag_SoftLepton_MuonTaggerNoIP_h
-
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-#include "RecoBTau/JetTagComputer/interface/JetTagComputer.h"
-#include "RecoBTag/SoftLepton/interface/LeptonSelector.h"
-#include "RecoBTag/SoftLepton/src/MuonTaggerNoIPMLP.h"
-
-/**  \class MuonTagger
- *
- *   Implementation of muon b-tagging using a softmax multilayer perceptron neural network
- *
- *
- *   \author Andrea 'fwyzard' Bocci, Universita' di Firenze
- */
-
-class MuonTaggerNoIP : public JetTagComputer {
-public:
-
-  /// explicit ctor
-  explicit MuonTaggerNoIP(const edm::ParameterSet & configuration) : 
-    m_selector(configuration)
-  { 
-    uses("slTagInfos"); 
-  }
-
-  /// dtor
-  virtual ~MuonTaggerNoIP() { }
-
-  /// b-tag a jet based on track-to-jet parameters in the extened info collection
-  virtual float discriminator(const TagInfoHelper & tagInfo) const;
-
-private:
-
-  btag::LeptonSelector m_selector;
-
-};
-
-#endif // RecoBTag_SoftLepton_MuonTaggerNoIP_h
